Started Animation Work
Created a separate scene for animation work, created animation controller with idle and jogging "Transitions" for a y bot model. Animations are in place and occur based on a getkey. An animation script was created for the player model in order to keep the movement script separate. The playerMove script moves the entire player object, whereas the animTest script resides in the ybot assets that is a child of the player object!

So far so good!

Next objective is to implement sprinting transition and jumping, and maybe climbing.