(Animation, Scripting, Importing)Jump,Run Back, and Transition
Imported a run back animation and deleted the old one, add the animation and tranisiton for the run back and jump animation. Jump animation makes the model ascend and not stay in place, but this works for now since the other  motions will have to be in place such as slide and climb.

Added code for animation transistions inside AnimTest.cs

Set backKey and jumpKey as "S" and "spacebar" respectively.