Started new Player Move, added assets to Leve Design1
We can't use the original player move since it clamps the direction of the jump, we need free movement during the jump.

Working on a new Player move script with the same player movement, but jumping will be more simpler.

Only created the same movement, jump is not implemented yet.

Added a plane and player object to Leve Design1

FPS scene/Player now has PlayerMove2 script.